pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Backend for Sora Reader"
authors = ["dhvcc <1337kwiz@gmail.com>"]
license = "GPLv3"

[tool.poetry.dependencies]
python = "^3.10"
Markdown = "^3.3.3"
psycopg2 = "^2.9.5"
Django = "^4.1.3"
django-extensions = "^3.2.1"
Werkzeug = "0.16.0"
django-environ = "^0.9.0"
dj-database-url = "^1.0.0"
django-cors-headers = "^3.13.0"
whitenoise = { extras = ["brotli"], version = "^6.2.0" }
django-jazzmin = "^2.6.0"
django-ninja = "^0.19.1"
django-ninja-jwt = "^5.2.2"
django-allauth = "^0.51.0"
django-rq = "^2.6.0"
rq-scheduler = "^0.11.0"
Scrapy = "^2.7.1"
scrapyscript = {git = "https://github.com/dhvcc/scrapyscript.git", rev = "error-handling"}
requests = "^2.25.1"
redis = "^4.3.4"
autoflake = "^1.7.7"
gunicorn = "^20.1.0"
orjson = "^3.8.1"
typesense = "^0.14.0"
colorlog = "^6.7.0"
sentry-sdk = "^1.11.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^5.0.4"
black = "^22.8.0"
isort = "^5.7.0"
pre-commit = "^2.11.1"
rich = "^12.5.1"
ipdb = "^0.13.9"
autoflake = "^1.6.0"
django-debug-toolbar = "^3.7.0"
pytest = "^7.2.0"
pytest-django = "^4.5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py310']
force-exclude = '''
(
    \.eggs
  | \.git
  | build
  | dist
  | venv
  | .venv
)
'''

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 100
skip_gitignore = true
skip_glob = ['**/.venv/**']
